* rename the binary to mylang * use anyhow crate for errors * move unit tests back to Rust idiomatic location in-file * refactor bulk of main to process module * add a process module for file and text processing * use fmt::display to print tokens
