+++ Cargo.toml
@@ -5,9 +5,10 @@ authors = ["Jonathan B Coe <jonathanbcoe@gmail.com>"]
 edition = "2018"
 
 [[bin]]
-name = "lexer"
+name = "mylang"
 path = "src/main.rs"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+anyhow = "1.0.41"
