--- src/lib.rs
@@ -1,7 +1,0 @@
-mod lexer;
-mod parser;
-mod token;
-
-pub use lexer::Lexer;
-pub use parser::Parser;
-pub use token::Kind;
